requirements.txt
fastapi~=0.104.1
uvicorn==0.15.0
python-jose==3.3.0
sqlalchemy~=2.0.0
python-multipart==0.0.6
PyJWT==2.8.0
hypercorn==0.14.4
Jinja2==3.1.2
websockets==12.0
databases==0.8.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic~=2.5.2
alembic~=1.13.0
requests~=2.31.0
asyncpg~=0.29.0
email-validator==2.1.0.post1
ElasticEmail~=4.0.10
fastapi-redis-cache==0.2.5
aioredis~=1.3.1
fastapi-cache2==0.2.1


pillow~=10.2.0
isbnlib~=3.10.14